Move node_modules to assets/js

+++ Makefile
@@ -1,23 +1,27 @@
 
+JS=_site/assets/js
+NM=_site/node_modules
 build: setup
 	bundler exec jekyll build
 	npx purgecss --css _site/assets/css/*.css --content `find _site -name "*.html" | grep -v _site/doc/ | grep -v _site/docs/`  -o _site/assets/css/
-	rm -f _site/assets/js/{bootstrap,@fortawesome,prismjs}
-	mkdir -p _site/assets/js/bootstrap/dist/js/ _site/assets/js/prismjs/components _site/assets/js/@fortawesome/fontawesome-free/
-	cp _site/node_modules/bootstrap/dist/js/bootstrap.*min.js* _site/assets/js/bootstrap/dist/js/
-	cp _site/node_modules/prismjs/prism.js _site/assets/js/prismjs/
-	cp _site/node_modules/prismjs/components/prism-erlang.js _site/assets/js/prismjs/components/
-	cp -r _site/node_modules/@fortawesome/fontawesome-free/webfonts _site/assets/js/@fortawesome/fontawesome-free/
-	rm -rf _site/node_modules/
+	# rm -rf $(JS)/{bootstrap,@fortawesome,prismjs}
+	# mkdir -p $(JS)/bootstrap/dist/js/ $(JS)/prismjs/components $(JS)/@fortawesome/fontawesome-free/
+	# cp $(NM)/bootstrap/dist/js/bootstrap.*min.js* $(JS)/bootstrap/dist/js/
+	# cp $(NM)/prismjs/prism.js $(JS)/prismjs/
+	# cp $(NM)/prismjs/components/prism-erlang.js $(JS)/prismjs/components/
+	# cp -r $(NM)/@fortawesome/fontawesome-free/webfonts $(JS)/@fortawesome/fontawesome-free/
+	# rm -rf $(NM)/
 
 vendor/bundle:
 	bundler install --path vendor/bundle
 
-node_modules: package-lock.json assets/js
+node_modules: package-lock.json
 	npm install
-	ln -s ../../node_modules/bootstrap assets/js/bootstrap
-	ln -s ../../node_modules/@fortawesome assets/js/@fortawesome
-	ln -s ../../node_modules/prismjs assets/js/prismjs
+
+setup_npm: node_modules assets/js
+	if [ ! -L assets/js/bootstrap ]; then ln -s ../../node_modules/bootstrap assets/js/bootstrap; fi
+	if [ ! -L assets/js/@fortawesome ]; then ln -s ../../node_modules/bootstrap assets/js/@fortawesome; fi
+	if [ ! -L assets/js/prismjs ]; then ln -s ../../node_modules/bootstrap assets/js/prismjs; fi
 
 otp_versions.table:
 	curl https://raw.githubusercontent.com/erlang/otp/master/otp_versions.table > $@
@@ -34,7 +38,7 @@ _data/releases.json: _releases node_modules _data/download.ts
 update:
 	npm update
 
-setup: vendor/bundle node_modules _data/releases.json docs
+setup: vendor/bundle setup_npm _data/releases.json docs
 
 serve: setup
 	bundle exec jekyll serve --trace --livereload
